refactor: EpicType type 폴더로 분리 [#100]
