+++ client/src/types/epic.d.ts
@@ -1,0 +1,6 @@
+export type EpicType = {
+  id: number;
+  name: string;
+  startAt: Date;
+  endAt: Date;
+};
